Move logic to a proper service
